chore: rename account-usage to usage
