+++ packages/jobs/package.json
@@ -21,7 +21,7 @@
         "@aws-sdk/client-s3": "3.993.0",
         "@aws-sdk/client-sqs": "3.993.0",
         "@kubernetes/client-node": "1.3.0",
-        "@nangohq/account-usage": "file:../account-usage",
+        "@nangohq/usage": "file:../usage",
         "@nangohq/data-ingestion": "file:../data-ingestion",
         "@nangohq/database": "file:../database",
         "@nangohq/fleet": "file:../fleet",
