+++ packages/lambda-runner/Dockerfile.lambda
@@ -42,7 +42,7 @@ COPY packages/providers/package.json ./packages/providers/package.json
 COPY packages/runner-sdk/package.json ./packages/runner-sdk/package.json
 COPY packages/billing/package.json ./packages/billing/package.json
 COPY packages/pubsub/package.json ./packages/pubsub/package.json
-COPY packages/account-usage/package.json ./packages/account-usage/package.json
+COPY packages/usage/package.json ./packages/usage/package.json
 COPY packages/email/package.json ./packages/email/package.json
 COPY packages/metering/package.json ./packages/metering/package.json
 COPY packages/lambda-runner/package.json ./packages/lambda-runner/package.json
