+++ packages/persist/tsconfig.json
@@ -27,7 +27,7 @@
             "path": "../billing"
         },
         {
-            "path": "../account-usage"
+            "path": "../usage"
         },
         {
             "path": "../pubsub"
