+++ packages/usage/lib/trace.ts
@@ -1,5 +1,5 @@
 import tracer from 'dd-trace';
 
 tracer.init({
-    service: 'nango-account-usage'
+    service: 'nango-usage'
 });
